% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexevent.R
\name{indexevent}
\alias{indexevent}
\title{Adjust association statistics for index event bias}
\usage{
indexevent(
  xbeta,
  xse,
  ybeta,
  yse,
  weighted = T,
  prune = NULL,
  method = c("CWLS", "Hedges-Olkin", "Simex"),
  tol = 1e-06,
  B = 10,
  lambda = seq(0.25, 5, 0.25),
  seed = 2018
)
}
\arguments{
\item{xbeta}{Vector of effects on the index trait}

\item{xse}{Vector of standard errors of \code{xbeta}}

\item{ybeta}{Vector of effects on the subsequent trait}

\item{yse}{Vector of standard errors of \code{ybeta}}

\item{weighted}{If true (default), regression of \code{ybeta} on \code{xbeta} is weighted by the inverse of \code{yse^2}.}

\item{prune}{Vector containing the indices of an approximately independent subset of the predictors in \code{xbeta} and \code{ybeta}.
If unspecified, all predictors will be used.}

\item{method}{Method to adjust for regression dilution (weak instruments) in the regression of \code{ybeta[prune]} on \code{xbeta[prune]}.
"CWLS" (default) applies Corrected Weighted Least Squares from Cai et al (2022).
"Hedges-Olkin" applies the correction from Dudbridge et al (2019), equivalent to CWLS for unweighted regression.
"Simex" applies a more time-consuming correction which may be more accurate than CWLS.}

\item{B}{Number of simulations performed in each stage of the Simex adjustment.}

\item{lambda}{Vector of lambdas for which the Simex simulations are performed.}

\item{seed}{Random number seed for the Simex adjustment}
}
\value{
An object of class "indexevent" which contains:
  \itemize{
    \item{\code{ybeta.adj} Adjusted effects on the subsequent trait}
    \item{\code{yse.adj} Adjusted standard errors of \code{ybeta.adj}}
    \item{\code{ychisq.adj} Chi-square statistics for \code{(ybeta.adj/yse.adj)^2}}
    \item{\code{yp.adj} P-values for \code{ychisq.adj} on 1df}
    \item{\code{b} Coefficient of the regression of \code{ybeta[prune]} on \code{xbeta[prune]}, after correction for regression dilution}
    \item{\code{b.se} Standard error of \code{b}}
    \item{\code{b.ci} Lower and upper confidence limits for \code{b}}
    \item{\code{b.raw} Regression coefficient without correction for regression dilution}
    \item{\code{simex.estimates} Regression coefficients under simulated measurement error}
  }
}
\description{
Given effect sizes and standard errors for predictors of an index trait and a subsequent trait,
this function adjusts the statistics for the subsequent trait for selection bias through the index trait.
}
\details{
Effect sizes are on a linear scale, so could be the coefficients from linear regression, or log odds ratios, or log hazard ratios.
Effects on the subsequent trait are regressed on the effects on the index trait.
By default, the regression is weighted by the inverse variances of the subsequent trait effects.
The regression is adjusted for sampling variation in the index trait effects,
and the residuals then used to obtain adjusted effect sizes and standard errors for the subsequent trait.

The regression should be performed on a subset of predictors that are independent.
In the context of a genome-wide association study, these would be LD-pruned SNPs.
In terms of the input parameters, the regression command is \code{lm(ybeta[prune]~xbeta[prune],weights=1/yse[prune]^2)}.

The effects in \code{xbeta} and \code{ybeta} should be aligned for the same variables
and the same direction prior to running \code{indexevent}.

The default value of \code{B} is 10 to get a quick result, but higher values are recommended, eg 1000.
}
\references{
Cai S, Hartley A, Mahmoud O, Tilling K, Dudbridge F (2022) Adjusting for collider bias in genetic association studies using instrumental variable methods. Genetic Epidemiol 46:303-316

Dudbridge F, Allen RJ, Sheehan NA, Schmidt AF, Lee JC, Jenkins RG, Wain LV, Hingorani AD, Patel RS (2019) Adjustment for index event bias in genome-wide association studies of subsequent events.  Nat Commun 10:1561
}
\author{
Frank Dudbridge
}
