% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simexllhd.R
\name{simexllhd}
\alias{simexllhd}
\title{Log-likelihood for SIMEX data}
\usage{
simexllhd(pvar, pmean, simex.estimates)
}
\arguments{
\item{pvar}{Ratio of the sampling variance to the variance of the true predictors, on the log scale.}

\item{pmean}{Slope of the simple linear regression.}

\item{simex.estimates}{Matrix containing data simulated by SIMEX.}
}
\value{
Log-likelihood evaluated at \code{pvar} and \code{pmean} for the data in \code{simex.estimates}.
}
\description{
Calculates the log-likelihood in a simple linear regression model
with measurement error in the predictor, using the SIMEX method.
}
\details{
\code{simex.estimates} is a matrix with three columns.
Column 1 contains the values of lambda under which measurement error is simulated.
Column 2 contains the estimated slopes for each value of lambda.
Column 3 contains the sampling variances of the estimated slopes.

The likelihood is a function of two parameters, the true slope of the simple linear regression and a
parameter representing the ratio of the sampling variance to the variance of the true predictors.
As this parameter must be positive, it is estimated on the log scale.
}
