% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simexprofileCI.R
\name{simexprofileCI}
\alias{simexprofileCI}
\title{Profile likelihood confidence interval for SIMEX}
\usage{
simexprofileCI(simex.estimates, variance.ratio)
}
\arguments{
\item{simex.estimates}{Matrix containing data simulated by SIMEX.}

\item{variance.ratio}{Ratio of the variance of the predictor to the variance of the outcome.}
}
\value{
A vector with three elements, the estimated slope and its lower and upper 95\% confidence limits.
}
\description{
Obtains a maximum likelihood estimate of the slope in a simple linear regression model
with measurement error in the predictor, using the SIMEX method.
}
\details{
\code{simex.estimates} is a matrix with three columns.
Column 1 contains the values of lambda under which measurement error is simulated.
Column 2 contains the estimated slopes for each value of lambda.
Column 3 contains the sampling variances of the estimated slopes.

The likelihood is a profile likelihood for the true regression slope, with the profile taken over a nuisance
parameter representing the ratio of the sampling variance to the variance of the true predictors.
The profiling step requires a value for the \code{variance.ratio}.
}
