% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simexprofilellhd.R
\name{simexprofilellhd}
\alias{simexprofilellhd}
\title{Profile log-likelihood for SIMEX data}
\usage{
simexprofilellhd(p, simex.estimates, variance.ratio)
}
\arguments{
\item{p}{Slope of the simple linear regression.}

\item{simex.estimates}{Matrix containing data simulated by SIMEX.}

\item{variance.ratio}{Ratio of the variance of the predictor to the variance of the outcome.}
}
\value{
Profile log-likelihood evaluated at \code{p} for the data in \code{simex.estimates}.
}
\description{
Calculates the profile log-likelihood of the slope in a simple linear regression model
with measurement error in the predictor, using the SIMEX method.
}
\details{
\code{simex.estimates} is a matrix with three columns.
Column 1 contains the values of lambda under which measurement error is simulated.
Column 2 contains the estimated slopes for each value of lambda.
Column 3 contains the sampling variances of the estimated slopes.

The likelihood is a profile likelihood for the true regression slope, with the profile taken over a nuisance
parameter representing the ratio of the sampling variance to the variance of the true predictors.
The lower bound for this nuisance parameter depends on \code{p} and \code{variance.ratio}.
}
