% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexevent.R
\name{indexevent}
\alias{indexevent}
\title{Adjust association statistics for index event bias}
\usage{
indexevent(xbeta, xse, ybeta, yse, prune, method = c("Hedges-Olkin", "Simex"),
  B = 1000, lambda = seq(0.25, 5, 0.25), seed = 2018)
}
\arguments{
\item{xbeta}{Vector of effects on the index trait}

\item{xse}{Vector of standard errors of \code{xbeta}}

\item{ybeta}{Vector of effects on the subsequent trait}

\item{yse}{Vector of standard errors of \code{ybeta}}

\item{prune}{Vector containing the indices of an approximately independent subset of the predictors in \code{xbeta} and \code{ybeta}}

\item{method}{Method to adjust for regression dilution in the regression of \code{ybeta[prune]} on \code{xbeta[prune]}.
"Hedges-Olkin" applies a quick but approximate correction.
"Simex" applies a more time-consuming, but accurate correction with proper allowance for its uncertainty.}

\item{B}{Number of simulations performed in each stage of the Simex adjustment}

\item{lambda}{Vector of lambdas for which the Simex simulations are performed.}

\item{seed}{Random number seed for the Simex adjustment}
}
\value{
An object of class "indexevent" which contains:
  \itemize{
    \item{\code{ybeta.adj} Adjusted effects on the subsequent trait}
    \item{\code{yse.adj} Adjusted standard errors of \code{ybeta.adj}}
    \item{\code{ychisq.adj} Chi-square statistics for \code{ybeta.adj/yse.adj}}
    \item{\code{yp.adj} P-values for \code{ychisq.adj} on 1df}
    \item{\code{b} Coefficient of the regression of \code{ybeta} on \code{xbeta}, after correction for regression dilution}
    \item{\code{b.se} Standard error of \code{b}}
    \item{\code{b.ci} Lower and upper confidence intervals for \code{b}}
    \item{\code{b.raw} Regression coefficient without correction for regression dilution}
    \item{\code{simex.estimates} Regression coefficients under simulated measurement error}
  }
}
\description{
Given effect sizes and standard errors for predictors of an index trait and a subsequent trait,
this function adjusts the statistics for the subsequent trait for selection bias through the index trait.
}
\details{
Effect sizes are on a linear scale, so could be the coefficients from linear regression, or log odds ratios, or log hazard ratios.
Effects on the subsequent trait are regressed on the effects on the index trait.
The regression is adjusted for sampling variation in the index trait effects,
and the residuals then used to obtain adjusted effect sizes and standard errors for the subsequent trait.
The regression should be performed on a subset of predictors that are independent.
In the context of a genome-wide association study, these would be LD-pruned SNPs.
}
\references{
Dudbridge F et al, Adjustment for index event bias in genome-wide association studies of subsequent events.  Submitted.
}
\author{
Frank Dudbridge
}
